@@ -4,7 +4,7 @@ description: A new Flutter project.
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.2.3 <4.0.0'
 
 dependencies:
   flutter:
@@ -19,7 +19,6 @@ dependencies:
   image_picker: ^1.0.4
   cupertino_icons: ^1.0.2
   json_annotation: ^4.7.0
-  flutter_form_builder: ^9.1.1
   intl_phone_field: ^3.1.0
   json_serializable: ^6.5.4
   shared_preferences: ^2.0.15
@@ -27,7 +26,8 @@ dependencies:
   built_collection: '>=5.1.1 <6.0.0'
   openapi:
     path: ./generated
-  flutter_lints: ^2.0.3
+  flutter_lints: ^3.0.1
+  flutter_form_builder: ^9.1.1
 
 
 
